ci: :sparkles: Created Attentance Card component
