+++ src/pages/hr/_config/columns/index.tsx
@@ -1,5 +1,6 @@
 import { IStatus } from '@/types';
 import { ColumnDef, Row } from '@tanstack/react-table';
+import { format } from 'date-fns';
 
 import PageAssign from '@/components/buttons/page-assign';
 import ResetPassword from '@/components/buttons/reset-password';
@@ -12,6 +13,8 @@ import { Switch } from '@/components/ui/switch';
 
 import { cn } from '@/lib/utils';
 
+import AttendanceCard from '../_components/attendance-card';
+import { formatDecimalHours } from '../utils';
 import {
 	DateAccessor,
 	DateAccessorDepartment,
@@ -25,7 +28,6 @@ import {
 	IMonthlyReportTableData,
 	IUserTableData,
 } from './columns.type';
-import { format } from 'date-fns';
 
 // Department Columns
 export const departmentColumns = (): ColumnDef<IDepartmentTableData>[] => [
@@ -449,14 +451,6 @@ export const deviceListColumns = ({
 ];
 
 //? Report ?//
-//* Format Decimal Values To Hours
-function formatDecimalHours(decimalHours: number) {
-	if (!decimalHours) return '--';
-	const totalMinutes = Math.round(decimalHours * 60);
-	const hours = Math.floor(totalMinutes / 60);
-	const minutes = totalMinutes % 60;
-	return `${hours}h ${minutes}m`;
-}
 
 //* Individual Report
 export const individualReportColumns = (dateAccessor: string[]): ColumnDef<IIndividualReportTableData>[] => [
@@ -495,104 +489,15 @@ export const individualReportColumns = (dateAccessor: string[]): ColumnDef<IIndi
 				info.getValue() as DateAccessor;
 
 			return (
-				<div
-					className={cn(
-						'w-[210px] space-y-3 rounded-lg border p-4 text-sm shadow-sm',
-						status === 'Absent' && 'border-red-200 bg-red-50',
-						status === 'Late' && 'border-yellow-200 bg-yellow-50',
-						status === 'Early Exit' && 'border-orange-200 bg-orange-50',
-						status === 'Present' && 'border-green-200 bg-green-50',
-						status === 'Off Day' && 'border-blue-200 bg-blue-50',
-						status === 'Leave' && 'border-purple-200 bg-purple-50'
-					)}
-				>
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Status:</strong>
-						<span
-							className={cn(
-								'rounded-full border px-3 py-0.5 text-xs font-medium capitalize',
-								status === 'Absent' && 'border-red-500 bg-red-100 text-red-700',
-								status === 'Late' && 'border-yellow-500 bg-yellow-100 text-yellow-700',
-								status === 'Early Exit' && 'border-orange-500 bg-orange-100 text-orange-700',
-								status === 'Present' && 'border-green-500 bg-green-100 text-green-700',
-								status === 'Off Day' && 'border-blue-500 bg-blue-100 text-blue-700',
-								status === 'Leave' && 'border-purple-500 bg-purple-100 text-purple-700'
-							)}
-						>
-							{status ?? 'N/A'}
-						</span>
-					</div>
-
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Entry:</strong>
-						{entry_time ? (
-							<DateTime
-								date={entry_time as Date}
-								isTime={true}
-								isDate={false}
-								ClassNameTime={cn(
-									'px-2 text-sm',
-									status === 'Late' && 'rounded-full bg-yellow-200 text-yellow-700'
-								)}
-							/>
-						) : (
-							<span className='px-2'>--</span>
-						)}
-					</div>
-
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Exit:</strong>
-						{exit_time ? (
-							<DateTime
-								date={exit_time as Date}
-								isTime={true}
-								isDate={false}
-								ClassNameTime={cn(
-									'px-2 text-sm',
-									status === 'Early Exit' && 'rounded-full bg-orange-200 text-yellow-700'
-								)}
-							/>
-						) : (
-							<span className='px-2'>--</span>
-						)}
-					</div>
-
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Late:</strong>
-						<span
-							className={cn(
-								'px-2',
-								status === 'Late' && 'rounded-full bg-yellow-200 px-2 text-yellow-700'
-							)}
-						>
-							{status === 'Absent' ? '-' : formatDecimalHours(late_hours as number)}
-						</span>
-					</div>
-
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Early Exit:</strong>
-						<span
-							className={cn(
-								'px-2',
-								status === 'Early Exit' && 'rounded-full bg-orange-200 px-2 text-orange-700'
-							)}
-						>
-							{status === 'Absent' ? '-' : formatDecimalHours(early_exit_hours)}
-						</span>
-					</div>
-
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Worked:</strong>
-						<span className='px-2'>
-							{status === 'Absent' ? '-' : formatDecimalHours(hours_worked as number)}
-						</span>
-					</div>
-
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Expected:</strong>
-						<span className='px-2'>{formatDecimalHours(expected_hours as number)}</span>
-					</div>
-				</div>
+				<AttendanceCard
+					status={status}
+					entry_time={entry_time}
+					exit_time={exit_time}
+					late_hours={late_hours as number}
+					early_exit_hours={early_exit_hours as number}
+					hours_worked={hours_worked as number}
+					expected_hours={expected_hours as number}
+				/>
 			);
 		},
 	})),
@@ -635,102 +540,15 @@ export const departmentReportColumns = (dateAccessor: string[]): ColumnDef<IDepa
 				info.getValue() as DateAccessorDepartment;
 
 			return (
-				<div
-					className={cn(
-						'w-[210px] space-y-3 rounded-lg border p-4 text-sm shadow-sm',
-						status === 'Absent' && 'border-red-200 bg-red-50',
-						status === 'Late' && 'border-yellow-200 bg-yellow-50',
-						status === 'Early Exit' && 'border-orange-200 bg-orange-50',
-						status === 'present' && 'border-green-200 bg-green-50',
-						status === 'Off Day' && 'border-blue-200 bg-blue-50',
-						status === 'Leave' && 'border-purple-200 bg-purple-50'
-					)}
-				>
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Status:</strong>
-						<span
-							className={cn(
-								'rounded-full border px-3 py-0.5 text-xs font-medium capitalize',
-								status === 'Absent' && 'border-red-500 bg-red-100 text-red-700',
-								status === 'Late' && 'border-yellow-500 bg-yellow-100 text-yellow-700',
-								status === 'Early Exit' && 'border-orange-500 bg-orange-100 text-orange-700',
-								status === 'present' && 'border-green-500 bg-green-100 text-green-700',
-								status === 'Off Day' && 'border-blue-500 bg-blue-100 text-blue-700',
-								status === 'Leave' && 'border-purple-500 bg-purple-100 text-purple-700'
-							)}
-						>
-							{status ?? 'N/A'}
-						</span>
-					</div>
-
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Entry:</strong>
-						{entry_time ? (
-							<DateTime
-								date={entry_time as Date}
-								isTime={true}
-								isDate={false}
-								ClassNameTime={cn(
-									'px-2 text-sm',
-									status === 'Late' && 'rounded-full bg-yellow-200 text-yellow-700'
-								)}
-							/>
-						) : (
-							<span className='px-2'>--</span>
-						)}
-					</div>
-
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Exit:</strong>
-						{exit_time ? (
-							<DateTime
-								date={exit_time as Date}
-								isTime={true}
-								isDate={false}
-								ClassNameTime={cn(
-									'px-2 text-sm',
-									status === 'Early Exit' && 'rounded-full bg-orange-200 text-yellow-700'
-								)}
-							/>
-						) : (
-							<span className='px-2'>--</span>
-						)}
-					</div>
-
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Late:</strong>
-						<span
-							className={cn(
-								'px-2',
-								status === 'Late' && 'rounded-full bg-yellow-200 px-2 text-yellow-700'
-							)}
-						>
-							{status === 'Absent' ? '-' : formatDecimalHours(late_hours)}
-						</span>
-					</div>
-
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Early Exit:</strong>
-						<span
-							className={cn(
-								'px-2',
-								status === 'Early Exit' && 'rounded-full bg-orange-200 px-2 text-orange-700'
-							)}
-						>
-							{status === 'Absent' ? '-' : formatDecimalHours(early_exit_hours)}
-						</span>
-					</div>
-
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Worked:</strong>
-						<span className='px-2'>{status === 'Absent' ? '-' : formatDecimalHours(hours_worked)}</span>
-					</div>
-
-					<div className='flex items-center gap-2'>
-						<strong className='min-w-[80px]'>Expected:</strong>
-						<span className='px-2'>{formatDecimalHours(expected_hours)}</span>
-					</div>
-				</div>
+				<AttendanceCard
+					status={status}
+					entry_time={entry_time}
+					exit_time={exit_time}
+					late_hours={late_hours as number}
+					early_exit_hours={early_exit_hours as number}
+					hours_worked={hours_worked as number}
+					expected_hours={expected_hours as number}
+				/>
 			);
 		},
 	})),
